Implement detached signal sender
Allow signals to be sent detached from the channel struct.

This is done using a higher-order function and by cloning the sender so
it can be used independently of the `Channel`.

This is useful to be able to spawn a cancelation handler separately from
the main `Channel` recieve loop.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>
f08f74

Signed-off-by: Joe Grund <jgrund@whamcloud.io>